+++ src/main/java/org/programmers/jpaweeklymission/customer/Infra/CustomerJpaRepository.java
@@ -1,5 +1,6 @@
-package org.programmers.jpaweeklymission.customer.data;
+package org.programmers.jpaweeklymission.customer.Infra;
 
+import org.programmers.jpaweeklymission.customer.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CustomerJpaRepository extends JpaRepository<Customer, Long> {
