+++ src/main/java/org/programmers/jpaweeklymission/customer/application/CustomerService.java
@@ -4,8 +4,8 @@ import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.programmers.jpaweeklymission.customer.presentation.dto.CustomerCreationRequest;
-import org.programmers.jpaweeklymission.customer.data.Customer;
-import org.programmers.jpaweeklymission.customer.data.CustomerJpaRepository;
+import org.programmers.jpaweeklymission.customer.Customer;
+import org.programmers.jpaweeklymission.customer.Infra.CustomerJpaRepository;
 import org.programmers.jpaweeklymission.customer.application.dto.CustomerResponse;
 import org.programmers.jpaweeklymission.customer.presentation.dto.CustomerUpdateRequest;
 import org.springframework.stereotype.Service;
